rename old -> legacy, new -> latest
